[add] cms trash can feature (#2220)
* [add] cms trash can feature

* [add] history trash task

* [modify] cms trash feature

* [fix] restore member photo file

* [modify] restore model relations

* [modify] restore model relation files

* [fix] 500 error when destroy history trash

* [fix] restore opendate resource file

* [modify] history trash parent

* [modify] search history trashes

* [fix] history trash lock_until

* [fix] restore cms page column_values

* [modify] history trashes spec

* [fix] restore ads banner file

* [fix] restore uploader file

* [fix] restore cms page without workflow

* [fix] restore opendata idea relations

* [fix] restore opendata app relations

* [fix] history initializer

+++ app/models/cms/column/value/file_upload.rb
@@ -138,9 +138,11 @@ class Cms::Column::Value::FileUpload < Cms::Column::Value::Base
 
   def destroy_file
     return if file.blank?
-
+    return nil unless File.exist?(file.path)
+    path = "#{Rails.root}/private/trash/#{file.path.sub(/.*\/(ss_files\/)/, '\\1')}"
+    FileUtils.mkdir_p(File.dirname(path))
+    FileUtils.cp(file.path, path)
     self.file.destroy
-    self.file_id = nil
   end
 
   # override Cms::Column::Value::Base#to_default_html
